#15 Followモデルのmodel specを実装する

+++ app/helpers/follows_helper.rb
@@ -1,5 +1,6 @@
 module FollowsHelper
   def follow_button(user)
+    return if current_user == user
     if Follow.exists?(following_id: current_user&.id, followed_id: user.id)
       button_to 'フォロー中', unfollow_path(user), method: :delete
     else
